Added all fns needed and changed comparison
There were some fns in the common_functions folder (commonFns) which needed to be hard written into folder. Now it should work independently.

Also I changed how the comparison works, so that you don't have to specify if it's phospho or not, just adds a count column of 1 for non-phospho